osmflatc: split into library + thin binary
This also adds a processing bench
